refactor(tui): move File tool tests to owned seam
Move the complete inline File tool test module into its own normally formatted test file. Production behavior and fixture bytes stay unchanged while the source-structure ratchet counts tests as tests.
